State_groupA y State_groupB --> State_Grupo
Se tenían dos estados separados segun el tipo de alimento (A o B) pero en ambos estados se hacía exactamente lo mismo (actGruposAlimentos()). Se han sustituido esos dos estados por un solo estado State_Grupo, ya que solo es necesario distinguir el tipo de alimento para saber el grupo exacto, pero eso ya se hace con los dos eventos TIPO_A y TIPO_B, así que no es necesario tener dos estados individuales.

Los cambios han afectado a las reglas de transición y las actividades de algunos estados, especialmente el STATE_ERROR.

+++ v2.1 - src/smartcloth_v2/Comida.h
@@ -37,7 +37,7 @@ class Comida{
 
   private:
   
-    int                   _nPlatos;                 /**< Número de platos en la comida. */
+    byte                  _nPlatos;                 /**< Número de platos en la comida. */
     float                 _peso;                    /**< Peso total de la comida. */
     ValoresNutricionales  _valoresComida;           /**< Valores nutricionales totales de la comida. */
 
@@ -51,14 +51,14 @@ class Comida{
      * 
      * @param num Número de platos a establecer
      */
-    inline void _setNumPlatos(int num){ _nPlatos = num; };
+    inline void _setNumPlatos(byte num){ _nPlatos = num; };
     
     /**
      * @brief Obtiene el número de platos de la comida.
      * 
      * @return El número de platos de la comida
      */
-    inline int _getNumPlatos(){ return _nPlatos; };
+    inline byte _getNumPlatos(){ return _nPlatos; };
